backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
web3
pydantic
